Remove dependency on Jest & migrate to Node.js test runner
