+++ __tests__/index.test.mjs
@@ -1,6 +1,11 @@
-const fs = require('fs');
-const path = require('path');
-const { ESLint } = require('eslint');
+import assert from 'node:assert/strict';
+import fs from 'node:fs';
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
+import { describe, it } from 'node:test';
+import { ESLint } from 'eslint';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const validExample = fs.readFileSync('./examples/valid.js', 'utf-8');
 const invalidExample = fs.readFileSync('./examples/invalid.js', 'utf-8');
@@ -13,14 +18,14 @@ describe('linting', () => {
 
   it('flags no warnings when valid', async () => {
     const lintResult = await eslint.lintText(validExample);
-    expect(lintResult[0].errorCount).toEqual(0);
-    expect(lintResult[0].warningCount).toEqual(0);
+    assert.strictEqual(lintResult[0].errorCount, 0);
+    assert.strictEqual(lintResult[0].warningCount, 0);
   });
 
   it('flags warnings when invalid', async () => {
     const [{ errorCount, warningCount }] =
       await eslint.lintText(invalidExample);
 
-    expect(errorCount + warningCount).toEqual(2);
+    assert.strictEqual(errorCount + warningCount, 2);
   });
 });
