+++ __tests__/rules.test.mjs
@@ -1,0 +1,26 @@
+import assert from 'node:assert/strict';
+import { describe, it } from 'node:test';
+import { ESLint } from 'eslint';
+
+import config from '../index.js';
+
+const getComputedConfig = async (baseConfig) => {
+  const eslint = new ESLint({
+    useEslintrc: false,
+    allowInlineConfig: false,
+    baseConfig,
+  });
+
+  const computedConfig = await eslint.calculateConfigForFile('index.js');
+  // Delete env-specific config keys.
+  delete computedConfig.filePath;
+  delete computedConfig.baseDirectory;
+
+  return computedConfig;
+};
+
+describe('rules snapshot', () => {
+  it('has stable config rules', async (t) => {
+    t.assert.snapshot(await getComputedConfig(config));
+  });
+});
