+++ package.json
@@ -28,7 +28,6 @@
     "eslint": "^8.56.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^9.1.0",
-    "jest": "^30.0.4",
     "prettier": "^3.6.2"
   },
   "peerDependencies": {
@@ -48,7 +47,7 @@
     "lint": "npm run lint:format && npm run lint:js && npm run lint:readme",
     "preflight": "npm run format && npm run lint && npm run test",
     "print-config": "eslint --print-config .eslintrc",
-    "test": "jest",
+    "test": "node --test",
     "write-rules": "node write-rules.js"
   },
   "prettier": {
